improved cloudflare worker adapter + added docs about different modes
