+++ app/src/adapter/index.ts
@@ -1,19 +1,14 @@
 import type { IncomingMessage } from "node:http";
 import type { App, CreateAppConfig } from "bknd";
 
-export type CfBkndModeCache<Env = any> = (env: Env) => {
-   cache: KVNamespace;
-   key: string;
-};
-
-export type CfBkndModeDurableObject<Env = any> = (env: Env) => {
-   durableObject: DurableObjectNamespace;
-   key: string;
-   keepAliveSeconds?: number;
-};
-
 export type CloudflareBkndConfig<Env = any> = {
-   mode?: CfBkndModeCache | CfBkndModeDurableObject;
+   mode?: "warm" | "fresh" | "cache" | "durable";
+   bindings?: (env: Env) => {
+      kv?: KVNamespace;
+      dobj?: DurableObjectNamespace;
+   };
+   key?: string;
+   keepAliveSeconds?: number;
    forceHttps?: boolean;
 };
 
@@ -27,14 +22,6 @@ export type BkndConfig<Env = any> = {
    };
    cloudflare?: CloudflareBkndConfig<Env>;
    onBuilt?: (app: App) => Promise<void>;
-};
-
-export type BkndConfigJson = {
-   app: CreateAppConfig;
-   setAdminHtml?: boolean;
-   server?: {
-      port?: number;
-   };
 };
 
 export function nodeRequestToRequest(req: IncomingMessage): Request {
